backend-services/requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.3.0
colorama==0.4.6
Django==4.2.6
django-cors-headers==4.3.0
django-filter==23.3
django-simple-history==3.4.0
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.0
filelock==3.12.4
future==0.18.3
gevent==23.9.1
greenlet==3.0.0
idna==3.4
Markdown==3.5
pydub==0.25.1
PyJWT==1.7.1
PySocks==1.7.1
pytube==15.0.0
pytz==2023.3.post1
requests==2.31.0
requests-file==1.5.1
six==1.16.0
sqlparse==0.4.4
tldextract==5.0.0
urllib3==2.0.6
zope.event==5.0
zope.interface==6.1
